+++ src/Factors/Mixture.jl
@@ -82,7 +82,8 @@ function getSample( cf::CalcFactor{<:Mixture},
   cf.factor.labels .= rand(cf.factor.diversity, N)
   for i in 1:N
     mixComponent = cf.factor.components[cf.factor.labels[i]]
-    smpls[1][i][:] = rand(mixComponent,1)
+    # measurements relate to the factor's manifold (either tangent vector or manifold point)
+    _setPointsMani!(smpls[1][i], randToPoints(mixComponent,1)[1])
   end
 
   # TODO only does first element of meas::Tuple at this stage, see #1099