+++ src/services/FactorGradients.jl
@@ -99,14 +99,6 @@ function FactorGradientsCached!(fct::Union{<:AbstractRelativeMinimize, <:Abstrac
                           λ_sizes )
 end
 
-_setPointsMani!(dest::AbstractArray, src::AbstractArray) = (dest .= src)
-
-function _setPointsMani!(dest::ProductRepr, src::ProductRepr)
-  for (k,prt) in enumerate(dest.parts)
-    _setPointsMani!(prt, src.parts[k])
-  end
-end
-
 
 # Gradient matrix has individual blocks
 getCoordSizes(fgc::FactorGradientsCached!) = fgc._coord_sizes