+++ NEWS.md
@@ -17,6 +17,8 @@ The list below highlights major breaking changes, and please note that significa
 - PPE methods used keyword `method::AbstractPointParametricType` which is now replaced with the keyword `ppeType`.
 - Belief points are now stored as a `Vector{P}` (rather than legacy `Matrix`), and currently still under the restriction `P <: AbstractVector{<:Real}`.  Objective is moving to `P` any user desired point that fits with the JuliaManifolds/Manifolds.jl patterns.
 - Deprecating use of `ensureAllInitialized!`, use `initAll!` instead.
+- New helper function `randToPoints(::SamplableBelief, N=1)::Vector{P}` to help with `getSample` for cases with new `ManifoldKernelDensity` beliefs for manifolds containing points of type `P`.
+
 # Major changes in v0.24
 
 - Update compat for ManifoldsBase.jl v0.11 with `AbstractManifold`.